Put every attack in a dedicated file and add tests
