+++ signature.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Signature:
+    h: int
+    r: int
+    s: int
